requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
httpx
requests
firebase-admin
gunicorn